HomeWork5, version2
