+++ src/main/java/com/samsung/domain/Question.java
@@ -1,10 +1,8 @@
 package com.samsung.domain;
 
-import lombok.AllArgsConstructor;
 import lombok.Value;
 
 
-@AllArgsConstructor
 @Value
 public class Question {
     int id;
